+++ docs/source/conf.py
@@ -26,7 +26,7 @@ copyright = "2022, Farshid Varno"
 year = "2022"
 author = "Farshid Varno"
 copyright = "{0}, {1}".format(year, author)
-version = release = "0.6.2"
+version = release = "0.7.0"
 # The full version, including alpha/beta/rc tags
 
 # -- General configuration ---------------------------------------------------
@@ -199,7 +199,7 @@ napoleon_use_ivar = True
 napoleon_use_rtype = False
 napoleon_use_param = False
 
-
+autodoc_inherit_docstrings = True
 # -----------------------------------------------------------------------------
 # Autosummary
 # -----------------------------------------------------------------------------