+++ fedsim/utils/convert_parameters.py
@@ -3,6 +3,10 @@ Parameters Conversion
 ---------------------
 """
 import torch
+from torch import Tensor
+from torch.nn import Module
+from torch.nn.utils import parameters_to_vector
+from torch.nn.utils import vector_to_parameters
 from torch.nn.utils.convert_parameters import _check_param_device
 
 
@@ -39,3 +43,44 @@ def vector_to_parameters_like(vec, parameters_like):
         # Increment the pointer
         pointer += num_param
     return new_params
+
+
+def vectorize_module(module: Module, clone=True, detach=True):
+    r"""convert parameters of a module to a vector
+
+    Args:
+        module (Module): module to convert the parameters of
+        clone (bool, optional): clones the output. Defaults to True.
+        detach (bool, optional): detaches the output. Defaults to True.
+
+    Returns:
+        Module: 1-D Tensor of all parameters in the module
+    """
+    vec = parameters_to_vector(module.parameters())
+    if clone:
+        vec = vec.clone()
+    if detach:
+        vec = vec.detach()
+    return vec
+
+
+def initialize_module(module: Module, vec: Tensor, clone=True, detach=True):
+    r"""initializes a module's parameters with a 1-D vector
+
+    Args:
+        module (Module): module to initialize weights
+        vec (Tensor): a 1-D Tensor
+        clone (bool, optional): clones the vector before initilization.
+            Defaults to True.
+        detach (bool, optional): detaches the output before the initialization.
+            Defaults to True.
+    """
+    if clone:
+        vec = vec.clone()
+    if detach:
+        vec = vec.detach()
+    if len(vectorize_module(module)) != len(vec):
+        return False
+
+    vector_to_parameters(vec, module.parameters())
+    return True